Major refactor in order to avoid n goroutines per service
Not all work is done, but we've actually laid most of the foundation.
Currently, the Governator type uses the Monitor type to listen for
SIGCHLD and avoids creating a goroutine per wait call. However,
os/exec still creates one goroutine per subprocess file descriptor
(one for stdout and one for stderr), so we'll have to avoid using it
in order to not create n goroutines per service.